# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication System API",
    "version": "1.0.0",
    "description": "API documentation for the Authentication System with Admin Roles"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/subscription/{id}": {
      "get": {
        "summary": "Get a subscription by ID",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription data"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      },
      "put": {
        "summary": "Update a subscription (Admin only)",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  },
                  "planId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription updated"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      },
      "delete": {
        "summary": "Cancel (delete) a subscription (Admin only)",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription cancelled"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
    "/api/subscription": {
      "post": {
        "summary": "Create a subscription",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "planId": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/subscription/user/{userId}": {
      "get": {
        "summary": "Get all subscriptions for a user",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user subscriptions"
          }
        }
      }
    },
    "/api/subscription/user/{userId}/stats": {
      "get": {
        "summary": "Get subscription statistics for a user",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription statistics"
          }
        }
      }
    },
    "/api/subscription/{id}/cancel": {
      "patch": {
        "summary": "Cancel a subscription",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription cancelled"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
    "/api/subscription/all": {
      "get": {
        "summary": "Get all subscriptions (Admin only)",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all subscriptions"
          }
        }
      }
    },
...
  },
...
}